Refactored AWS CLI commands into separate files
