+++ packages/cli/src/index.ts
@@ -1,8 +1,8 @@
-import { MedplumClient, normalizeErrorString, MEDPLUM_VERSION } from '@medplum/core';
+import { MEDPLUM_VERSION, MedplumClient, normalizeErrorString } from '@medplum/core';
 import { Command } from 'commander';
 import dotenv from 'dotenv';
 import { login, whoami } from './auth';
-import { aws } from './aws';
+import { aws } from './aws/index';
 import { bot, createBotDeprecate, deployBotDeprecate, saveBotDeprecate } from './bots';
 import { project } from './projects';
 import { deleteObject, get, patch, post, put } from './rest';
